LayoutTests/imported/w3c/web-platform-tests/wasm/jsapi/table/set-reftypes.tentative.any.worker-expected.txt

CONSOLE MESSAGE: NetworkError: Response is not 2xx

Harness Error (FAIL), message = NetworkError: Response is not 2xx
